Use HashCode as GroupID (and removed some warnings)
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@18217 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 2a43e064b47782a172300edeb9f8d6e0a8404c77

+++ server/core/core/src/de/uniol/inf/is/odysseus/core/server/physicaloperator/aggregate/IGroupProcessor.java
@@ -22,14 +22,14 @@ import de.uniol.inf.is.odysseus.core.server.physicaloperator.aggregate.basefunct
 
 public interface IGroupProcessor<R, W extends IClone> {
 
-	public Integer getGroupID(R elem);
+	public Long getGroupID(R elem);
 
 	public void init();
 
-	public W createOutputElement(Integer groupID,
+	public W createOutputElement(Long groupID,
 			PairMap<SDFSchema, AggregateFunction, W, ?> r);
 
-	public W createOutputElement2(Integer groupID,
+	public W createOutputElement2(Long groupID,
 			PairMap<SDFSchema, AggregateFunction, IPartialAggregate<R>, ?> e);
 
 }
